% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_technoise.R
\name{add_expr_noise}
\alias{add_expr_noise}
\title{Add experimental noise to true counts}
\usage{
add_expr_noise(results, ...)
}
\arguments{
\item{results}{The scMultisim result object}

\item{...}{\code{randseed}: The random seed
\code{protocol}: \code{UMI} or \code{non-UMI}
\code{gene_len}:  A vector with lengths of all genes
\code{alpha_mean}, \code{alpha_sd}: rate of subsampling of transcripts during capture step
\code{depth_mean}, \code{depth_sd}: The sequencing depth}
}
\value{
none
}
\description{
Add experimental noise to true counts
}
\examples{
results <- sim_example(ncells = 10)
add_expr_noise(results)
}
\seealso{
The underlying methods
\link{True2ObservedCounts} and \link{True2ObservedATAC}
}
