% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_technoise.R
\name{add_outliers}
\alias{add_outliers}
\title{Add outliers to the observed counts}
\usage{
add_outliers(
  res,
  prob = 0.01,
  factor = 2,
  sd = 0.5,
  cell.num = 1,
  max.var = Inf
)
}
\arguments{
\item{res}{The scMultisim result object}

\item{prob}{The probability of adding outliers for each gene}

\item{factor}{The factor of the outliers}

\item{sd}{The standard deviation of the outliers}

\item{cell.num}{For a gene, the number of cells chosen to add outliers}

\item{max.var}{The maximum variance allowed}
}
\value{
none
}
\description{
Add outliers to the observed counts
}
