% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{ATAC_SEQ}
\alias{ATAC_SEQ}
\title{Getting ATAC SEQ data}
\usage{
ATAC_SEQ(evf_ATAC, region_effect, dens_nonzero, p_zero, randseed)
}
\arguments{
\item{evf_ATAC}{EVF values for the ATAC features}

\item{region_effect}{like the gene_effect for the ATAC-Seq feature}

\item{dens_nonzero}{this is the density function of log(x+1), where x is the non-zero values for ATAC-SEQ data}

\item{p_zero}{the proportion of 0s we see in the ATAC-seq data}

\item{randseed}{should produce same result if all other parameters are all the same}
}
\value{
ATAC SEQ data
}
\description{
Getting ATAC SEQ data
}
