% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_functions.R
\name{cal_AUC}
\alias{cal_AUC}
\title{calculate area under curve}
\usage{
cal_AUC(x_vec, y_vec)
}
\arguments{
\item{x_vec}{vector of x values}

\item{y_vec}{vector of y values}
}
\description{
calculate area under curve
}
