% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_functions.R
\name{cal_KNNP}
\alias{cal_KNNP}
\title{calculate knn purity of a data space with given labels}
\usage{
cal_KNNP(data, label)
}
\arguments{
\item{data}{input data}

\item{label}{known label of the cells}
}
\description{
calculate knn purity of a data space with given labels
}
