% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dens_nonzero}
\alias{dens_nonzero}
\title{this is the density function of log(x+1), where x is the non-zero values for ATAC-SEQ data}
\format{
a vector.
}
\usage{
data(dens_nonzero)
}
\description{
this is the density function of log(x+1), where x is the non-zero values for ATAC-SEQ data
}
\examples{
data(dens_nonzero)
}
\keyword{datasets}
