% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_technoise.R
\name{divide_batches}
\alias{divide_batches}
\title{Divide batches for observed counts}
\usage{
divide_batches(results, nbatch = 2, effect = 3)
}
\arguments{
\item{results}{The scMultisim result object, after running \code{add_expr_noise}}

\item{nbatch}{Number of batches}

\item{effect}{Batch effect size, default is 3}
}
\description{
Divide batches for observed counts
}
\seealso{
The underlying method \link{DivideBatches}
}
