% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_main.R
\name{.continuous_cif}
\alias{.continuous_cif}
\title{Generates cifs for cells sampled along the trajectory of cell development}
\usage{
.continuous_cif(
  seed,
  N,
  options,
  ncell_key = "cell",
  is_spatial = F,
  spatial_params = NULL,
  .plot = F,
  .plot.name = "cont_cif.pdf"
)
}
\arguments{
\item{N}{the number list}

\item{options}{the option list}

\item{is_spatial}{return a list of cifs for spatial}
}
\description{
Generates cifs for cells sampled along the trajectory of cell development
}
