% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_technoise.R
\name{.divideBatchesImpl}
\alias{.divideBatchesImpl}
\title{Divide the observed counts into multiple batches by adding batch effect to each batch}
\usage{
.divideBatchesImpl(
  counts,
  meta_cell,
  nbatch,
  batch_effect_size = 1,
  randseed = 0
)
}
\arguments{
\item{counts}{gene cell matrix}

\item{meta_cell}{the meta information related to cells, will be combined with technical cell level information and returned}

\item{nbatch}{number of batches}

\item{batch_effect_size}{amount of batch effects. Larger values result in bigger differences between batches. Default is 1.}

\item{randseed}{random seed}
}
\description{
Divide the observed counts into multiple batches by adding batch effect to each batch
}
