% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_technoise.R
\name{.expandToBinary}
\alias{.expandToBinary}
\title{expand transcript counts to a vector of binaries of the same length of as the number of transcripts}
\usage{
.expandToBinary(true_counts_1cell)
}
\arguments{
\item{true_counts_1cell}{number of transcript in one cell}
}
\description{
expand transcript counts to a vector of binaries of the same length of as the number of transcripts
}
