% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_main.R
\name{.gene_identify_vectors}
\alias{.gene_identify_vectors}
\title{Title}
\usage{
.gene_identify_vectors(seed, sim, options)
}
\arguments{
\item{options}{}
}
\value{
a list of kon, koff (gene x cif+regu), s (gene x cif)
}
\description{
Title
}
