% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_main.R
\name{.get_numbers}
\alias{.get_numbers}
\title{Title}
\usage{
.get_numbers(GRN, options)
}
\arguments{
\item{options}{}
}
\value{
A list with the following keys
gene
grn.gene
non.grn.gene
regulator
target
cif
diff.cif
nd.cif
}
\description{
Title
}
