% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_benchmark.R
\name{.getCountCorrMatrix}
\alias{.getCountCorrMatrix}
\title{This function finds the correlation between every pair of genes}
\usage{
.getCountCorrMatrix(counts)
}
\arguments{
\item{counts}{rna seq counts}
}
\value{
the correlation matrix
}
\description{
This function finds the correlation between every pair of genes
}
