% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_sim.R
\name{.getParams}
\alias{.getParams}
\title{Get Kineic Parameters for all cells and genes}
\usage{
.getParams(seed, sim, sp_cell_i = NULL, sp_path_i = NULL)
}
\arguments{
\item{seed}{random seed}

\item{sim}{the simulation environment}

\item{sp_cell_i}{spatial cell index}

\item{sp_path_i}{the pre-sampled path along the tree for this cell}
}
\value{
the kinetic parameters
}
\description{
Get Kineic Parameters for all cells and genes
}
