% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_main.R
\name{.normalize_GRN_params}
\alias{.normalize_GRN_params}
\title{Rename the original gene IDs in the GRN table to integers.}
\usage{
.normalize_GRN_params(params)
}
\value{
list
}
\description{
Rename the original gene IDs in the GRN table to integers.
}
