% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_main.R
\name{.normalizeGRNParams}
\alias{.normalizeGRNParams}
\title{Rename the original gene IDs in the GRN table to integers.}
\usage{
.normalizeGRNParams(params)
}
\arguments{
\item{params}{GRN parameters.}
}
\value{
list
}
\description{
Rename the original gene IDs in the GRN table to integers.
}
