% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_main.R
\name{.region_identity_vectors}
\alias{.region_identity_vectors}
\title{Title}
\usage{
.region_identity_vectors(seed, GRN, N, options)
}
\arguments{
\item{options}{}
}
\value{
region x cif matrix
}
\description{
Title
}
