% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_technoise.R
\name{.rnormTrunc}
\alias{.rnormTrunc}
\title{sample from truncated normal distribution}
\usage{
.rnormTrunc(n, mean, sd, a, b)
}
\arguments{
\item{n}{number of values to create}

\item{mean}{mean of the normal distribution}

\item{sd}{standard deviation of the normal distribution}

\item{a}{the minimum value allowed}

\item{b}{the maximum value allowed}
}
\value{
a vector of length n
}
\description{
sample from truncated normal distribution
}
