% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.1_spatial.R
\name{gen_clutter}
\alias{gen_clutter}
\title{generate a clutter of cells by growing from the center}
\usage{
gen_clutter(
  n_cell,
  grid_size = NA,
  center = c(0, 0),
  existing_loc = NULL,
  existing_grid = NULL
)
}
\arguments{
\item{n_cell}{the number of cells}

\item{grid_size}{the width and height of the grid}

\item{center}{the center of the grid}

\item{existing_loc}{only place cells on the specified existing locations}

\item{existing_grid}{manually specify what locations are in the grid}
}
\value{
a matrix of locations
}
\description{
generate a clutter of cells by growing from the center
}
\examples{
gen_clutter(10, 10, c(5, 5))

}
