% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_benchmark.R
\name{gene_corr_regulator}
\alias{gene_corr_regulator}
\title{Print the correlations between targets of each regulator}
\usage{
gene_corr_regulator(results = .getResultsFromGlobal(), regulator)
}
\arguments{
\item{results}{The scMultisim result object}

\item{regulator}{The regulator ID in the GRN params}
}
\value{
none
}
\description{
Print the correlations between targets of each regulator
}
\examples{
results <- simExample200Cells()
gene_corr_regulator(results)
}
