% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{GeneEffects}
\alias{GeneEffects}
\title{Getting GeneEffects matrices}
\usage{
GeneEffects(
  num_genes,
  num_evfs,
  randseed,
  prob,
  geffect_mean,
  geffect_sd,
  num_diffEVF,
  regulator_ID_list,
  target_gene_ID_list,
  gene_effects_by_regulator
)
}
\arguments{
\item{num_genes}{number of genes}

\item{num_evfs}{number of evfs}

\item{randseed}{should produce same result if all other parameters are all the same}

\item{prob}{the probability that the effect size is not 0}

\item{geffect_mean}{the mean of the normal distribution where the non-zero effect sizes are dropped from}

\item{geffect_sd}{the standard deviation of the normal distribution where the non-zero effect sizes are dropped from}

\item{num_diffEVF}{number of diff EVFs}

\item{regulator_ID_list}{list of the IDs of regulator genes}

\item{target_gene_ID_list}{list of the IDs of target genes}

\item{gene_effects_by_regulator}{gene effects of each regulator gene (column) on every gene ID (row)}
}
\value{
a list of 3 matrices, each of dimension ngenes * nevf
}
\description{
This function randomly generates the effect size of each evf on the dynamic expression parameters
}
