% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_benchmark.R
\name{Get_ATAC_correlation}
\alias{Get_ATAC_correlation}
\title{This function gets the average correlation rna seq counts and chromatin region effect on genes}
\usage{
Get_ATAC_correlation(counts, atacseq_data, num_genes)
}
\arguments{
\item{counts}{rna seq counts}

\item{atacseq_data}{atac seq data}

\item{num_genes}{number of genes}
}
\value{
the correlation value
}
\description{
This function gets the average correlation rna seq counts and chromatin region effect on genes
}
\examples{
results <- sim_example(ncells = 10)
Get_ATAC_correlation(results$counts, results$atacseq_data, results$num_genes)
}
