% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{get_prob}
\alias{get_prob}
\title{Get probability based on gene length}
\usage{
get_prob(glength)
}
\arguments{
\item{glength}{the the gene length}
}
\value{
prob, the probability
}
\description{
Get probability based on gene length
}
