% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_functions.R
\name{GetCountCorrMatrix}
\alias{GetCountCorrMatrix}
\title{This function finds the correlation between every pair of genes}
\usage{
GetCountCorrMatrix(counts)
}
\arguments{
\item{counts}{rna seq counts}
}
\description{
This function finds the correlation between every pair of genes
}
