% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_functions.R
\name{GetGeneModuleColors}
\alias{GetGeneModuleColors}
\title{This function assigns each gene to a gene module based on which regulator gene has the largest effect on the gene}
\usage{
GetGeneModuleColors(
  GRN_params,
  gene_effects_by_regulator,
  num_genes,
  randseed = 0
)
}
\arguments{
\item{GRN_params}{GRN_params is a matrix where: #    - column 1 is the target gene ID, #    - column 2 is the gene ID which acts as a transcription factor for the target (regulated) gene #    - column 3 is the effect of the column 2 gene ID on the column 1 gene ID}

\item{gene_effects_by_regulator}{gene effects of each regulator gene (column) on every gene ID (row)}

\item{num_genes}{number of genes}

\item{randseed}{should produce same result if all other parameters are all the same}
}
\description{
This function assigns each gene to a gene module based on which regulator gene has the largest effect on the gene
}
