% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_functions.R
\name{GetGeneModuleCorrelation}
\alias{GetGeneModuleCorrelation}
\title{This function gets the average correlation between genes which are most co-regulated by a regulator gene}
\usage{
GetGeneModuleCorrelation(
  counts,
  GRN_params,
  gene_effects_by_regulator,
  num_genes
)
}
\arguments{
\item{counts}{rna seq counts}

\item{GRN_params}{GRN_params is a matrix where: #    - column 1 is the target gene ID, #    - column 2 is the gene ID which acts as a transcription factor for the target (regulated) gene #    - column 3 is the effect of the column 2 gene ID on the column 1 gene ID}

\item{gene_effects_by_regulator}{gene effects of each regulator gene (column) on every gene ID (row)}

\item{num_genes}{number of genes}
}
\description{
This function gets the average correlation between genes which are most co-regulated by a regulator gene
}
