% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{loadRData}
\alias{loadRData}
\title{loads an RData file, and returns it}
\usage{
loadRData(fileName)
}
\arguments{
\item{fileName}{RData file name}
}
\description{
loads an RData file, and returns it
}
