% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{LogDist}
\alias{LogDist}
\title{Getting logged expression distribution}
\usage{
LogDist(counts, log_count_bins)
}
\arguments{
\item{counts}{the expression matrix}

\item{log_count_bins}{a vector of the cut-offs for the histogram}
}
\value{
a matrix where the rows are the genes and columns are the number of samples within a count category
}
\description{
Prepares for plotting the Count Heatmap
}
