% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_opts.R
\name{OP}
\alias{OP}
\title{Get option from an object in the current environment}
\usage{
OP(..., .name = "options")
}
\arguments{
\item{...}{the parameter name}

\item{.name}{get option from this object}
}
\value{
the parameter value
}
\description{
Get option from an object in the current environment
}
