% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_utils.R
\name{Phyla1}
\alias{Phyla1}
\title{Creating a linear example tree}
\usage{
Phyla1(len = 1)
}
\arguments{
\item{len}{length of the tree}
}
\value{
a R phylo object
}
\description{
Creating a linear example tree
}
\examples{
Phyla1(len = 1)
}
