% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_utils.R
\name{Phyla5}
\alias{Phyla5}
\title{Creating an example tree with 5 tips}
\usage{
Phyla5(plotting = F)
}
\arguments{
\item{plotting}{True for plotting the tree on console, False for no plot}
}
\value{
a R phylo object
}
\description{
Creating an example tree with 5 tips
}
