% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_benchmark.R
\name{plot_cell_loc}
\alias{plot_cell_loc}
\title{Plot cell locations}
\usage{
plot_cell_loc(
  results = .getResultsFromGlobal(),
  size = 4,
  show.label = F,
  show.arrows = T,
  lr.pair = 1,
  .cell.pop = NULL
)
}
\arguments{
\item{results}{The scMultisim result object}

\item{size}{Fig size}

\item{show.label}{Show cell numbers}

\item{show.arrows}{Show arrows representing cell-cell interactions}

\item{lr.pair}{The ligand-receptor pair used to plot CCI arrows
\code{results$cci_cell_type_param[lr.pair]}}

\item{.cell.pop}{Specify the cell population metadata}
}
\value{
none
}
\description{
Plot cell locations
}
\examples{
results <- simExample200CellsSpatial()
plot_cell_loc(results)
}
