% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_benchmark.R
\name{plot_gene_module_cor_heatmap}
\alias{plot_gene_module_cor_heatmap}
\title{Plot the gene module correlation heatmap}
\usage{
plot_gene_module_cor_heatmap(
  results = .get_results_from_global(),
  seed = 0,
  grn.genes.only = T,
  save = F
)
}
\arguments{
\item{results}{The scMultisim result object}

\item{seed}{The random seed}

\item{grn.genes.only}{Plot the GRN gens only}

\item{save}{}
}
\description{
Plot the gene module correlation heatmap
}
