% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_benchmark.R
\name{plot_grid}
\alias{plot_grid}
\title{Plot the CCI grid}
\usage{
plot_grid(results = .get_results_from_global())
}
\arguments{
\item{results}{The scMultisim result object}
}
\description{
In normal cases, please use \code{plot_cell_loc} instead.
}
