% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_benchmark.R
\name{plot_grn}
\alias{plot_grn}
\title{Plot the GRN network}
\usage{
plot_grn(params)
}
\arguments{
\item{params}{The GRN params data frame}
}
\value{
none
}
\description{
Plot the GRN network
}
\examples{
data(GRN_params_100, envir = environment())
plot_grn(GRN_params_100)
}
