% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_benchmark.R
\name{plot_phyla}
\alias{plot_phyla}
\title{Plot a R phylogenic tree}
\usage{
plot_phyla(tree)
}
\arguments{
\item{tree}{The tree}
}
\value{
none
}
\description{
Plot a R phylogenic tree
}
\examples{
plotPhyla(Phyla5())
}
