% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_benchmark.R
\name{plot_tsne}
\alias{plot_tsne}
\title{Plot a (D x cells) data matrix using TSNE.}
\usage{
plot_tsne(
  data,
  labels,
  perplexity = 60,
  legend = "",
  plot.name = "",
  save = F,
  rand.seed = 0
)
}
\arguments{
\item{data}{}

\item{perplexity}{}

\item{rand.seed}{}

\item{label}{}

\item{saving}{}

\item{plotname}{}
}
\description{
Plot a (D x cells) data matrix using TSNE.
}
