% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_functions.R
\name{PlotTsne}
\alias{PlotTsne}
\title{Perform tSNE}
\usage{
PlotTsne(
  meta,
  data,
  perplexity = 60,
  label,
  saving = T,
  plotname,
  randseed = 0
)
}
\arguments{
\item{meta}{simulation parameters}

\item{data}{expression matrix}

\item{perplexity}{the perplexity parameter used in the tsne visualization}

\item{label}{the column name of the meta data that the points needs to be colored by}

\item{saving}{whether or not to save the plot}

\item{plotname}{the name of the jpeg file}

\item{randseed}{should produce same result if all other parameters are all the same}
}
\description{
This function takes the ground truth (meta), the expresssion matrix (data), and performs tSNE, saves the result to a jpeg file with user specified name
}
