% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{Region2Gene}
\alias{Region2Gene}
\title{Getting Region2Gene matrices}
\usage{
Region2Gene(num_genes, nregions, cumulative_probability, randseed)
}
\arguments{
\item{num_genes}{number of genes}

\item{nregions}{regions 1 to nregion are considered sequentially located on the genome}

\item{cumulative_probability}{the probability that a gene is regulated by respectively 0, 1, 2 consecutive regions}

\item{randseed}{should produce same result if all other parameters are all the same}
}
\value{
region2gene, of shape (nregions) x (num_genes), indicates which genes are associated with which regions
}
\description{
Getting Region2Gene matrices
}
