% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{RegionEffects}
\alias{RegionEffects}
\title{Getting RegionEffect matrix}
\usage{
RegionEffects(nregions, nevf, randseed, prob, effect_mean, effect_sd)
}
\arguments{
\item{nregions}{number of regions}

\item{nevf}{number of evfs}

\item{randseed}{(should produce same result if nregions, nevf and randseed are all the same)}

\item{prob}{the probability that the effect size is not 0}

\item{effect_mean}{the mean of the normal distribution where the non-zero effect sizes are dropped from}

\item{effect_sd}{the standard deviation of the normal distribution where the non-zero effect sizes are dropped from}
}
\value{
a matrix of dimension nregions * nevf
}
\description{
Getting RegionEffect matrix
}
