% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{RegulatorGeneEffects}
\alias{RegulatorGeneEffects}
\title{Getting RegulatorGeneEffects matrices}
\usage{
RegulatorGeneEffects(
  num_genes,
  randseed,
  GRN_params,
  regulator_ID_list,
  target_gene_ID_list
)
}
\arguments{
\item{num_genes}{number of genes}

\item{randseed}{should produce same result if all other parameters are all the same}

\item{GRN_params}{GRN_params is a matrix where: #    - column 1 is the target gene ID, #    - column 2 is the gene ID which acts as a transcription factor for the target (regulated) gene #    - column 3 is the effect of the column 2 gene ID on the column 1 gene ID}

\item{regulator_ID_list}{list of the IDs of regulator genes}

\item{target_gene_ID_list}{list of the IDs of target genes}
}
\value{
a list of 3 matrices, each of dimension ngenes * nevf
}
\description{
Getting RegulatorGeneEffects matrices
}
