% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{RNA_SEQ}
\alias{RNA_SEQ}
\title{Getting RNA_SEQ true counts by cell one by one taking into account the GRN conditions of the previous cell}
\usage{
RNA_SEQ(
  evf_all,
  phyla,
  num_genes,
  num_cells,
  num_regulators,
  evf_center,
  gene_effects,
  gene_effects_by_regulator,
  match_params_den,
  params,
  intrinsic_noise,
  regulator_ID_list,
  scale_s,
  Sigma,
  do_velocity,
  beta,
  d,
  num_cycles,
  cycle_length
)
}
\arguments{
\item{evf_all}{a list of two object, one is the evf, and the other is a dataframe indicating the branch each cell comes from (pop) and its depth in the tree}

\item{phyla}{the cell developmental tree if chosing 'discrete' or 'continuous' evf type. Can either be generated randomly (using pbtree(nclusters) function from phytools package) or read from newick format file using the ape package}

\item{num_genes}{number of genes}

\item{num_cells}{number of cells}

\item{num_regulators}{number of regulator genes}

\item{evf_center}{mean value of normal distribution used to generate evf start}

\item{gene_effects}{a list of three matrices (generated using the GeneEffects function),}

\item{gene_effects_by_regulator}{gene effects of each regulator gene (column) on every gene ID (row)}

\item{match_params_den}{density parameters that data is matched to}

\item{params}{the parameters for simulating gene expression from EVf and gene effects}

\item{intrinsic_noise}{a 0 to 1 value which is the weight assigned to the random sample from the Beta-Poisson distribution, where the weight of the Beta-Poisson mean value is given a weight of 1 minus the intrinsic noise value.}

\item{regulator_ID_list}{list of the IDs of regulator genes}

\item{scale_s}{a factor to scale the s parameter, which is used to tune the size of the actual cell (small cells have less number of transcripts in total)}

\item{Sigma}{parameter of the std of evf values within the same population}

\item{do_velocity}{set to TRUE to incorporate the RNA velocity simulation module to simulate unspliced and spliced RNA sequence counts and ground truth velocity}

\item{beta}{splicing rate of each gene}

\item{d}{degradation rate of each gene}

\item{num_cycles}{for generating velocity data, the number of cycles run before sampling the gene expression of a cell}

\item{cycle_length}{for generating velocity data, a factor which is multiplied by the expected time to transition from kon to koff and back to to form the the length of a cycle}
}
\value{
generate a true RNA SEQ count matrix, which has the shape (genes) x (cells)
}
\description{
Getting RNA_SEQ true counts by cell one by one taking into account the GRN conditions of the previous cell
}
