% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_utils.R
\name{SampleDen}
\alias{SampleDen}
\title{sample from smoothed density function}
\usage{
SampleDen(nsample, den_fun)
}
\arguments{
\item{nsample}{number of samples needed}

\item{den_fun}{density function estimated from density() from R default}
}
\value{
a vector of samples
}
\description{
sample from smoothed density function
}
