% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9_meta.R
\name{scmultisim_help}
\alias{scmultisim_help}
\title{Show detailed documentations of scMultiSim's parameters}
\usage{
scmultisim_help(topic = NULL)
}
\arguments{
\item{topic}{Can be \code{options}, \code{dynamic.GRN}, or \code{cci}}
}
\description{
Show detailed documentations of scMultiSim's parameters
}
\examples{
scmultisim_help()
}
