% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_functions.R
\name{sens_and_spec}
\alias{sens_and_spec}
\title{calculate sensitivity and specificity}
\usage{
sens_and_spec(isDE_gold, isDE_pred)
}
\arguments{
\item{isDE_gold}{true values}

\item{isDE_pred}{predicted values}
}
\description{
calculate sensitivity and specificity
}
