% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_utils.R
\name{sim_example}
\alias{sim_example}
\title{Simulate a small example dataset with 200 cells and the 100-gene GRN}
\usage{
sim_example(ncells = 10, velocity = FALSE)
}
\arguments{
\item{ncells}{number of cells, please increase this number on your machine}

\item{velocity}{whether to simulate RNA velocity}
}
\value{
the simulation result
}
\description{
Simulate a small example dataset with 200 cells and the 100-gene GRN
}
\examples{
sim_example(ncells = 10)
}
