% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_utils.R
\name{sim_example_200_cells}
\alias{sim_example_200_cells}
\title{Simulate a small example dataset with 200 cells and the 100-gene GRN}
\usage{
sim_example_200_cells(velocity = FALSE)
}
\arguments{
\item{velocity}{whether to simulate RNA velocity}
}
\value{
the simulation result
}
\description{
Simulate a small example dataset with 200 cells and the 100-gene GRN
}
\examples{
sim_example_200_cells()
}
