% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_utils.R
\name{sim_example_200_cells_spatial}
\alias{sim_example_200_cells_spatial}
\title{Simulate a small example dataset with 200 cells and the 100-gene GRN, with CCI enabled}
\usage{
sim_example_200_cells_spatial()
}
\value{
the simulation result
}
\description{
Simulate a small example dataset with 200 cells and the 100-gene GRN, with CCI enabled
}
