% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_utils.R
\name{sim_example_spatial}
\alias{sim_example_spatial}
\title{Simulate a small example dataset with 200 cells and the 100-gene GRN, with CCI enabled}
\usage{
sim_example_spatial(ncells = 10)
}
\arguments{
\item{ncells}{number of cells, please increase this number on your machine}
}
\value{
the simulation result
}
\description{
Simulate a small example dataset with 200 cells and the 100-gene GRN, with CCI enabled
}
\examples{
sim_example_spatial(ncells = 10)
}
