% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_spatial_functions.R
\name{SimulateTrueCounts_spatial}
\alias{SimulateTrueCounts_spatial}
\title{Generate true results according to simulation parameters}
\usage{
SimulateTrueCounts_spatial(
  GRN_params,
  num_cells = 1000,
  unregulated_to_regulated_gene_ratio = 0.1,
  num_evfs = 500,
  diffEVF_fraction = 0.9,
  Sigma = 0.1,
  atac_effect = 0.5,
  beta = 0.4,
  d = 1,
  num_cycles = 3,
  cycle_length = 1,
  intrinsic_noise = 1,
  randseed = 0,
  do_velocity = FALSE,
  phyla = Phyla5(),
  nregions_distribution = cumsum(c(0.1, 0.5, 0.4)),
  p_zero = 0.8,
  vary = "s",
  evf_center = 1,
  impulse = F,
  bimod = 0,
  geffect_mean = 0,
  gene_effect_prob = 0.3,
  gene_effects_sd = 1,
  reffect_mean = 0,
  region_effect_prob = 0.3,
  region_effects_sd = 1,
  scale_s = 1,
  dyn_grn_params = NULL,
  spatial_params = NULL,
  mg_kon = NULL
)
}
\arguments{
\item{GRN_params}{GRN_params is a matrix where: #    - column 1 is the target gene ID, #    - column 2 is the gene ID which acts as a transcription factor for the target (regulated) gene #    - column 3 is the effect of the column 2 gene ID on the column 1 gene ID}

\item{num_cells}{total number of cells from all populations}

\item{unregulated_to_regulated_gene_ratio}{ratio of unregulated genes to regulated genes.  The final number of genes is rounded up to the nearest 10}

\item{num_evfs}{number of EVFs for each kinetic parameter}

\item{diffEVF_fraction}{fraction of evfs which are differential evfs between cell types}

\item{Sigma}{parameter of the std of evf values within the same cell type, controls heterogeneity for each cell type}

\item{atac_effect}{a 0 to 1 value which is the weight assigned to the influence of chromatin accessability data on gene expression}

\item{beta}{splicing rate of each gene}

\item{d}{degradation rate of each gene}

\item{num_cycles}{for generating velocity data, the number of cycles run before sampling the gene expression of a cell}

\item{cycle_length}{for generating velocity data, a factor which is multiplied by the expected time to transition from kon to koff and back to to form the the length of a cycle}

\item{intrinsic_noise}{a 0 to 1 value which is the weight assigned to the random sample from the Beta-Poisson distribution, where the weight of the Beta-Poisson mean value is given a weight of 1 minus the intrinsic noise value.}

\item{randseed}{should produce same result if all other parameters are all the same}

\item{do_velocity}{set to TRUE to incorporate the RNA velocity simulation module to simulate unspliced and spliced RNA sequence counts and ground truth velocity}

\item{phyla}{a tree which defines relationship between populations}

\item{nregions_distribution}{the probability that a gene is regulated by respectively 0, 1, ..., (length(nregions_distribution) - 1) consecutive regions}

\item{p_zero}{the proportion of 0s we see in the ATAC-seq data}

\item{vary}{which kinetic parameters have differential evfs. Can be "all", "kon", "koff", "s", "except_kon", "except_koff", "except_s"}

\item{evf_center}{the value which evf mean is generated from (default=1)}

\item{impulse}{when generating continuous populations, use the impulse model or not. Default is FALSE}

\item{bimod}{adjusts the bimodality of gene expression, thus controlling intrinsic variation}

\item{geffect_mean}{the mean of gene effect size}

\item{gene_effect_prob}{probability of non-zero values in the gene effect vectors}

\item{gene_effects_sd}{controls differences between genes}

\item{reffect_mean}{the mean of region effect size}

\item{region_effect_prob}{probability of non-zero values in the region effect vectors}

\item{region_effects_sd}{controls differences between regions}

\item{scale_s}{the cell size parameter in (0,1). Use smaller value for cell types known to be small (like naive cells)}
}
\value{
a list of eleven elements: 1. A matrix containing the true (spliced) transcript counts, 2. Gene level meta information, 3. Cell level meta information, including a matrix of EVFs and a vector of cell identity (for example, the population the cell belongs to), 4. The parameters k~on~, k~off~ and s used to simulation the true counts, 5. A matrix containing the true unspliced transcript counts, 6. The true RNA velocity information for each cell, 7. The pseudotime at which the cell counts were generated, 8. The scATAC-seq data, 9. The matrix region2gene, a 0 1 matrix of shape (nregions) x (num_genes), where a value of 1 indicates the gene is affected by the accessibility of the particular chromatin region, 10. Gene effects of each regulator gene (column) on every gene ID (row), 11. The number of genes used in the experiment
}
\description{
Generate true results according to simulation parameters
}
