% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.1_spatial.R
\docType{class}
\name{spatialGrid-class}
\alias{spatialGrid-class}
\alias{.SpatialGrid}
\title{The class for spatial grids}
\description{
The class for spatial grids
}
\section{Fields}{

\describe{
\item{\code{method}}{the method to generate the cell layout}

\item{\code{grid_size}}{the width and height of the grid}

\item{\code{ncells}}{the number of cells}

\item{\code{grid}}{the grid matrix}

\item{\code{locs}}{a list containing the locations of all cells}

\item{\code{loc_order}}{deprecated, don't use; the order of the locations}

\item{\code{cell_types}}{a map to save the cell type of each allocated cell}

\item{\code{same_type_prob}}{the probability of a new cell placed next to a cell with the same type}

\item{\code{max_nbs}}{the maximum number of neighbors for each cell}

\item{\code{nb_map}}{a list containing the neighbors for each cell}

\item{\code{nb_adj}}{adjacency matrix for neighbors}

\item{\code{nb_radius}}{the radius of neighbors}

\item{\code{final_types}}{the final cell types after the final time step}

\item{\code{pre_allocated_pos}}{the pre-allocated positions for each cell, if any}

\item{\code{method_param}}{additional parameters for the layout method}
}}


