% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_technoise.R
\name{divide_batches}
\alias{divide_batches}
\title{Divide batches for observed counts}
\usage{
divide_batches(results, nbatch = 2, effect = 3, randseed = 0)
}
\arguments{
\item{results}{The scMultisim result object, after running \code{addExprNoise()}}

\item{nbatch}{Number of batches}

\item{effect}{Batch effect size, default is 3}

\item{randseed}{Random seed}
}
\value{
none
}
\description{
Divide batches for observed counts
}
\examples{
results <- sim_example_200_cells()
add_expr_noise(results)
divide_batches(results)
}
