% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_benchmark.R
\name{gene_corr_cci}
\alias{gene_corr_cci}
\title{Plot the ligand-receptor correlation summary}
\usage{
gene_corr_cci(
  results = .get_results_from_global(),
  all.genes = F,
  .pair = NULL,
  .exclude.same.types = T
)
}
\arguments{
\item{results}{The scMultisim result object}
}
\description{
Plot the ligand-receptor correlation summary
}
