% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_benchmark.R
\name{gene_corr_cci}
\alias{gene_corr_cci}
\title{Plot the ligand-receptor correlation summary}
\usage{
gene_corr_cci(
  results = .getResultsFromGlobal(),
  all.genes = F,
  .pair = NULL,
  .exclude.same.types = T
)
}
\arguments{
\item{results}{The scMultisim result object}

\item{all.genes}{Whether to use all genes or only the ligand/receptor genes}

\item{.pair}{Return the raw data for the given LR pair}

\item{.exclude.same.types}{Whether to exclude neighbor cells with same cell type}
}
\value{
none
}
\description{
Plot the ligand-receptor correlation summary
}
