% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_benchmark.R
\name{plot_grid}
\alias{plot_grid}
\title{Plot the CCI grid}
\usage{
plot_grid(results = .getResultsFromGlobal())
}
\arguments{
\item{results}{The scMultisim result object}
}
\value{
none
}
\description{
In normal cases, please use \code{plotCellLoc} instead.
}
\examples{
results <- sim_example_200_cells_spatial()
plot_grid(results)
}
